requirements.txt
fastapi
uvicorn
python-multipart
openai
transformers
datasets
torch
pytesseract
pdfminer
python-docx
pdfminer.six
Pillow
accelerate>=0.26.0
nltk
together
PyPDF2